+++ test/clojure_lsp/features/call_hierarchy_test.clj
@@ -35,7 +35,9 @@
         "(defn d-func []"
         "  (println 1))"))
 
-(def project-root-uri "file:/")
+(def core-code
+  (code "(ns clojure.core)"
+        "(defn println [& args])"))
 
 (deftest prepare
   (h/load-code-and-locs a-code (h/file-uri "file:///some/a.clj"))
@@ -43,7 +45,7 @@
   (h/load-code-and-locs c-code (h/file-uri "file:///some/c.clj"))
   (h/load-code-and-locs d-code (h/file-uri "file:///some/d.clj"))
   (testing "single element"
-    (let [items (f.call-hierarchy/prepare (h/file-uri "file:///some/d.clj") 2 7 project-root-uri)]
+    (let [items (f.call-hierarchy/prepare (h/file-uri "file:///some/d.clj") 2 7)]
       (is (= 1 (count items)))
       (is (= {:name            "d-func []"
               :kind            :function
@@ -69,7 +71,7 @@
                :uri (h/file-uri "file:///some/c.clj")
                :range {:start {:line 2 :character 6} :end {:line 2 :character 12}}
                :selection-range {:start {:line 2 :character 6} :end {:line 2 :character 12}}}}]
-      (f.call-hierarchy/incoming (h/file-uri "file:///some/d.clj") 2 7 project-root-uri)))
+      (f.call-hierarchy/incoming (h/file-uri "file:///some/d.clj") 2 7)))
 
   (testing "for multiple elements"
     (h/assert-submaps
@@ -97,9 +99,10 @@
                :uri (h/file-uri "file:///some/b.clj")
                :range {:start {:line 5 :character 6} :end {:line 5 :character 14}}
                :selection-range {:start {:line 5 :character 6} :end {:line 5 :character 14}}}}]
-      (f.call-hierarchy/incoming (h/file-uri "file:///some/c.clj") 3 7 project-root-uri))))
+      (f.call-hierarchy/incoming (h/file-uri "file:///some/c.clj") 3 7))))
 
 (deftest outgoing
+  (h/load-code-and-locs core-code (h/file-uri "jar:file:///.m2/clojure.jar!/clojure/core.clj"))
   (h/load-code-and-locs a-code (h/file-uri "file:///some/a.clj"))
   (h/load-code-and-locs b-code (h/file-uri "file:///some/b.clj"))
   (h/load-code-and-locs c-code (h/file-uri "file:///some/c.clj"))
@@ -115,7 +118,7 @@
              :uri (h/file-uri "file:///some/b.clj")
              :range {:start {:line 2 :character 6} :end {:line 2 :character 12}}
              :selection-range {:start {:line 2 :character 6} :end {:line 2 :character 12}}}}]
-      (f.call-hierarchy/outgoing (h/file-uri "file:///some/a.clj") 3 7 project-root-uri)))
+      (f.call-hierarchy/outgoing (h/file-uri "file:///some/a.clj") 3 7)))
 
   (testing "for multiple elements"
     (h/assert-submaps
@@ -135,4 +138,16 @@
              :uri (h/file-uri "file:///some/c.clj")
              :range {:start {:line 2 :character 6} :end {:line 2 :character 12}}
              :selection-range {:start {:line 2 :character 6} :end {:line 2 :character 12}}}}]
-      (f.call-hierarchy/outgoing (h/file-uri "file:///some/b.clj") 3 7 project-root-uri))))
+      (f.call-hierarchy/outgoing (h/file-uri "file:///some/b.clj") 3 7)))
+
+  (testing "with external elements"
+    (h/assert-submaps
+      [{:from-ranges []
+        :to {:name "println [& args]"
+             :kind :variable
+             :tags []
+             :detail "clojure.core"
+             :uri "zipfile:///.m2/clojure.jar::clojure/core.clj"
+             :range {:start {:line 1 :character 6} :end {:line 1 :character 13}}
+             :selection-range {:start {:line 1 :character 6} :end {:line 1 :character 13}}}}]
+      (f.call-hierarchy/outgoing (h/file-uri "file:///some/d.clj") 2 7))))